refactor: Split code from transfer.py into auth.py sdk_wrappers.py modules
Split up the transfer.py module into a few different modules to better
organize the code.
